requirements.txt
groq==0.18.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-experimental==0.3.4
langchain-groq==0.2.4
numpy==1.26.4
pandas==2.2.3
plotly==6.0.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
seaborn==0.13.2
streamlit==1.42.2
tabulate==0.9.0

